rename: station DTO 패키지 분리
