+++ src/main/java/subway/station/StationController.java
@@ -11,6 +11,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import subway.station.dto.StationRequest;
+import subway.station.dto.StationResponse;
+
 @RestController
 public class StationController {
     private StationService stationService;
